Добавление связи индикатора и дисциплины, тестирование и исправление

+++ app/UKUP/moduleDB.py
@@ -47,6 +47,7 @@ def get_disciplines(direction, year):
         .join(DirectionDiscipline, Discipline.id == DirectionDiscipline.discipline_id) \
         .filter(DirectionDiscipline.direction_id == direction) \
         .filter(Discipline.year_approved == year) \
+        .order_by(Discipline.module_id) \
         .all()
     return disciplines
 
@@ -187,7 +188,7 @@ def get_competences(direction, year):
     competences = Competence.query \
         .filter_by(direction_id=direction) \
         .filter_by(year_approved=year) \
-        .all()
+        .order_by(Competence.type.desc()).order_by(Competence.name).all()
     return competences
 
 
@@ -211,8 +212,10 @@ def add_competence(competence_params):
 
 
 # Функция привязки дисциплины к компетенциям
-def update_discipline_competences(discipline_id, competence_ids):
-    existing_links = CompetenceDiscipline.query.filter_by(discipline_id=discipline_id).all()
+def update_discipline_competences(discipline_id, competence_ids, direction_id):
+    existing_links = CompetenceDiscipline.query.filter_by(discipline_id=discipline_id).join(Competence) \
+        .filter(Competence.direction_id == direction_id) \
+        .all()
     existing_competence_ids = {link.competence_id for link in existing_links}
 
     # Ищем компетенции, которые нужно добавить
@@ -278,18 +281,26 @@ def update_discipline_indicators(discipline_id, competence_ids, indicator_ids):
 def delete_competence(id_competence):
     # Находим компетенцию по ее идентификатору
     competence = Competence.query.get(id_competence)
+    if competence is None:
+        raise ValueError("Компетенция с идентификатором {} не найдена".format(id_competence))
 
     # Удаляем все связи с дисциплинами для этой компетенции
     CompetenceDiscipline.query.filter_by(competence_id=id_competence).delete()
 
     # Находим все связи с индикаторами для этой компетенции
     indicator_links = IndicatorDiscipline.query \
+        .join(Indicator, Indicator.id == IndicatorDiscipline.indicator_id) \
         .filter(Indicator.competence_id == id_competence) \
         .all()
 
-    # Удаляем все связи с индикаторами для этой компетенции
+    indicators = Indicator.query.filter(Indicator.competence_id == id_competence).all()
+
+    # Удаляем все связи с индикаторами и дисциплинами для этой компетенции
     for link in indicator_links:
         db.session.delete(link)
+
+    for indicator in indicators:
+        db.session.delete(indicator)
 
     # Удаляем саму компетенцию
     db.session.delete(competence)
@@ -423,7 +434,7 @@ def get_indicators_disciplines_links_by_competence_id(competence_id):
 
     # Получаем список связей дисциплин и индикаторов
     links = db.session.query(IndicatorDiscipline)\
-                      .join(IndicatorDiscipline.indicator)\
+                      .join(Indicator, IndicatorDiscipline.indicator_id == Indicator.id)\
                       .filter_by(competence_id=competence_id)\
                       .all()
 
@@ -489,10 +500,13 @@ def delete_connection(connection_id):
 
 
 # Функция получения компетенций, привязанных к дисциплине
-def get_connected_competences(discipline_id):
-    connected_competences = Competence.query.join(CompetenceDiscipline, CompetenceDiscipline.competence_id == Competence.id)\
-                                            .filter(CompetenceDiscipline.discipline_id == discipline_id)\
-                                            .all()
+def get_connected_competences(discipline_id, direction_id, year):
+    connected_competences = Competence.query \
+        .join(CompetenceDiscipline, CompetenceDiscipline.competence_id == Competence.id) \
+        .filter(CompetenceDiscipline.discipline_id == discipline_id) \
+        .filter(Competence.direction_id == direction_id) \
+        .filter(Competence.year_approved == year) \
+        .all()
     return connected_competences
